changes of borrow book, return book and some changes in delete methods of books, authors and users

+++ OnlineBookStore.Bussiness/Service/AuthorService.cs
@@ -1,4 +1,6 @@
-﻿using Microsoft.EntityFrameworkCore;
+﻿using Microsoft.AspNetCore.Http;
+using Microsoft.EntityFrameworkCore;
+using OnlineBookStore.Business.Services;
 using OnlineBookStore.Bussiness.IService;
 using OnlineBookStore.Bussiness.ViewModels;
 using OnlineBookStore.Bussiness.ViewModels.Author;
@@ -9,27 +11,58 @@ using System.Net;
 
 namespace OnlineBookStore.Bussiness.Service
 {
-    public class AuthorService : IAuthorService
+    public class AuthorService : BaseService, IAuthorService
     {
         private readonly ApplicationDbContext _context;
 
-        public AuthorService(ApplicationDbContext context)
+        public AuthorService(ApplicationDbContext context, IHttpContextAccessor httpContextAccessor) : base(httpContextAccessor)
         {
             _context = context;
         }
 
-        public async Task<ApiResponse> GetAllAuthorsAsync()
+        public async Task<ApiResponse> GetAllAuthorsAsync(AuthorFilterRequest authorFilter)
         {
-            var authors = await _context.Authors
-                .Select(a => new AuthorResponse
+            var query = _context.Authors.AsQueryable();
+
+            if (!string.IsNullOrWhiteSpace(authorFilter.Name))
+            {
+                query = query.Where(a => a.Name.ToLower().Contains(authorFilter.Name.ToLower()));
+            }
+
+            var sortBy = authorFilter.SortBy?.ToLower() ?? "name";
+            var sortOrder = authorFilter.SortOrder?.ToLower() ?? "asc";
+
+            query = sortBy switch
+            {
+                "bio" => sortOrder == "desc" ? query.OrderByDescending(a => a.Bio) : query.OrderBy(a => a.Bio),
+                _ => sortOrder == "desc" ? query.OrderByDescending(a => a.Name) : query.OrderBy(a => a.Name),
+            };
+
+            var pageNumber = authorFilter.PageNumber <= 0 ? 1 : authorFilter.PageNumber;
+            var pageSize = authorFilter.PageSize <= 0 ? 10 : authorFilter.PageSize;
+
+            var totalRecords = await query.CountAsync();
+
+            var authors = await query
+                .Skip((pageNumber - 1) * pageSize)
+                .Take(pageSize)
+                .Select(a => new
                 {
-                    Id = a.Id,
-                    Name = a.Name,
-                    Bio = a.Bio
+                    a.Id,
+                    a.Name,
+                    a.Bio
                 })
                 .ToListAsync();
 
-            return new ApiResponse(HttpStatusCode.OK, new List<string> { string.Format(CommonMessage.DataFetched, "Authors") }, authors);
+            var responseData = new
+            {
+                TotalRecords = totalRecords,
+                PageNumber = pageNumber,
+                PageSize = pageSize,
+                Authors = authors
+            };
+
+            return new ApiResponse(HttpStatusCode.OK, new List<string> { CommonMessage.DataFetched.Replace("{0}", "Authors") }, responseData);
         }
 
         public async Task<ApiResponse> GetAuthorByIdAsync(int id)
@@ -94,15 +127,35 @@ namespace OnlineBookStore.Bussiness.Service
 
         public async Task<ApiResponse> DeleteAuthorAsync(int id)
         {
-            var author = await _context.Authors.FindAsync(id);
+            var author = await _context.Authors
+                .Include(a => a.Books)
+                .FirstOrDefaultAsync(a => a.Id == id);
 
             if (author == null)
-                return new ApiResponse(HttpStatusCode.NotFound, new List<string> { string.Format(CommonMessage.NotFound, "Author") });
+            {
+                return new ApiResponse(HttpStatusCode.NotFound, new List<string>
+                {
+                    string.Format(CommonMessage.NotFound, "Author")
+                });
+            }
+
+            if (author.Books != null && author.Books.Any())
+            {
+                var bookTitles = author.Books.Select(b => b.Title).ToList();
+                var booksList = string.Join(", ", bookTitles);
+
+                string message = string.Format(CommonMessage.CannotDeleteAuthorWithBooks, booksList);
+
+                return new ApiResponse(HttpStatusCode.BadRequest, new List<string> { message });
+            }
 
             _context.Authors.Remove(author);
             await _context.SaveChangesAsync();
 
-            return new ApiResponse(HttpStatusCode.OK, new List<string> { string.Format(CommonMessage.Deleted, "Author") });
+            return new ApiResponse(HttpStatusCode.OK, new List<string>
+            {
+                string.Format(CommonMessage.Deleted, "Author")
+            });
         }
     }
 }
